requirements.txt
Django==5.1.6
djangorestframework==3.15.2
python-dotenv==1.0.1
djangorestframework-simplejwt==5.4.0
django-filter==25.1
celery==5.4.0
redis==5.2.1
python-dotenv==1.0.1
drf-yasg==1.21.8
pytest-django==4.10.0
pytest-cov==6.0.0
pytest==8.3.4
